# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BotalSePaisa Operations Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginCredentials {
  username: string;
  password: string;
}

export interface AuthToken {
  token: string;
  username: string;
  role: string;
}

export interface Shop {
  id: number;
  shopId: string;
  shopName: string;
  ownerName: string;
  mobile: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  locationCapturedAt?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface ShopInput {
  /** @minLength 1 */
  shopName: string;
  /** @minLength 1 */
  ownerName: string;
  /** @minLength 1 */
  mobile: string;
  latitude?: number;
  longitude?: number;
  accuracy?: number;
  photoBase64: string;
}

export interface ShopUpdate {
  /** @minLength 1 */
  shopName?: string;
  /** @minLength 1 */
  ownerName?: string;
  /** @minLength 1 */
  mobile?: string;
  latitude?: number;
  longitude?: number;
  accuracy?: number;
  photoBase64?: string;
}

export interface ShopList {
  data: Shop[];
  total: number;
  page: number;
  limit: number;
}

export interface Collection {
  id: number;
  shopId: number;
  shop?: Shop;
  collectionDate: string;
  weightKg: number;
  ratePerKg: number;
  totalAmount: number;
  paymentStatus: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  paidBy?: string | null;
  createdAt: string;
}

export interface CollectionInput {
  shopId: number;
  collectionDate: string;
  weightKg: number;
}

export interface CollectionList {
  data: Collection[];
  total: number;
  page: number;
  limit: number;
}

export interface PaymentUpdate {
  paymentStatus: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  paidBy?: string | null;
}

export interface ReportEntry {
  id: number;
  shopId: number;
  shop: Shop;
  collectionDate: string;
  weightKg: number;
  ratePerKg: number;
  totalAmount: number;
  paymentStatus: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  paidBy?: string | null;
  createdAt: string;
}

export interface ReportResult {
  data: ReportEntry[];
  totalKg: number;
  totalAmount: number;
}

export interface Settings {
  id: number;
  pricePerKg: number;
}

export interface SettingsUpdate {
  pricePerKg: number;
}

export interface DashboardSummary {
  totalShops: number;
  totalKgCollected: number;
  totalAmountPayable: number;
  pendingPayments: number;
  recentCollections: Collection[];
}

export type ListShopsParams = {
/**
 * Search by shop name, owner name, or mobile
 */
search?: string;
/**
 * Page number
 */
page?: number;
/**
 * Items per page
 */
limit?: number;
};

export type ListCollectionsParams = {
shopId?: number;
startDate?: string;
endDate?: string;
paymentStatus?: string;
page?: number;
limit?: number;
};

export type GetReportsParams = {
shopId?: number;
/**
 * ISO week string e.g. 2024-W01
 */
week?: string;
/**
 * Month string e.g. 2024-01
 */
month?: string;
startDate?: string;
endDate?: string;
};