# lib/api-zod/src/generated/types/shop.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BotalSePaisa Operations Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Shop {
  id: number;
  shopId: string;
  shopName: string;
  ownerName: string;
  mobile: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  locationCapturedAt?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}