# lib/api-zod/src/generated/types/shopUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BotalSePaisa Operations Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ShopUpdate {
  /** @minLength 1 */
  shopName?: string;
  /** @minLength 1 */
  ownerName?: string;
  /** @minLength 1 */
  mobile?: string;
  latitude?: number;
  longitude?: number;
  accuracy?: number;
  photoBase64?: string;
}